requirements.txt
folium==0.17.0
geojson==3.1.0
gpsd-py3==0.3.0
tile-operator==0.0.6